+++ CHANGELOG.md
@@ -2,11 +2,32 @@
 
 ## master
 
+- Add `Process.project` and `rubanok_scope` methods to get the Hash of recognized params. ([@palkan][])
+
+```ruby
+class PostsProcessor < Rubanok::Processor
+  map :q { ... }
+  match :page, :per_page, activate_on: :page { ... }
+end
+
+PostsProcessor.project(q: "search_me", filter: "smth", page: 2)
+# => { q: "search_me", page: 2 }
+
+class PostsController < ApplicationController
+  def index
+    @filter_params = rubanok_scope
+    # or
+    @filter_params = rubanok_scope params.require(:filter), with: PostsProcessor
+    # ...
+  end
+end
+```
+
 - Improve naming by using "processor" instead of "plane". ([@palkan][])
 
-  See [the discussion](https://github.com/palkan/rubanok/issues/3).
+See [the discussion](https://github.com/palkan/rubanok/issues/3).
 
-  **NOTE**: Older API is still available without deprecation.
+**NOTE**: Older API is still available without deprecation.
 
 - Add `fail_when_no_matches` parameter to `match` method. ([@Earendil95][])